requirements.txt
streamlit == 1.31.1
pandas == 2.0.3
seaborn == 0.12.2
matplotlib == 3.7.2
numpy == 1.24.3
scikit-learn == 1.3.0
plotly == 5.9.0
Pillow == 10.0.1